+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added `thiserror` and `log` crates
 - Updated `Trade` and `ExecutedTrade` to use better nomenclature for trade types.
 - Change default threshold for `BBands`
+- Add logging for when outside of risk tolerance
 
 ### Portfolio Improvements